generalize typeclass constraints slightly
With these changes we can do

    allExpressions $ map mkCon [4, 6, 3 :: Integer]

to make expressions with Integer constants, or

    map Expr.evalExpr $ allExpressions $ map mkCon [4, 6, 3 :: Integer] :: [Rational]

to evaluate those expressions to rationals.

The cost is I no longer understand how to write `fromRational` instances
because

   fromRational :: (Eq a, Real a) => Fractional (Ezpr a)
   fromRational r = ???

I can't simply use `Con r` here because Con has to be able to produce
Ezpr a, and `Con r` is Ezpr Rational.  I need a function that turns
Rational -> a.  fromRational is almost that: Fractional a => Rational
-> a.  But the whole point here is that I don't want to require
fromRational a.  If there were

   numerator, denominator :: Num a => Rational -> a

then I could do it.