Get rid of annoying dependence on GHC.Real
I said a couple of commits ago that I wanted numerator and denominator functions.
Data.Ratio has them.

Also add fromRational definition for Exprs, using the same technique,
and loosen the constraints required to infer Fractional (Ezpr a):
we don't need Real a, just Num a.  (Real a is left over from when I thought
I might use toRational, but that died on the vine.)